fix : reviewed comments changes
