+++ CSharp_Assignments.sln
@@ -48,7 +48,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "CircularDependencies", "Cir
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Assignment19", "src\Assignment19\Assignment19.csproj", "{492A5075-1101-4095-A16B-44670F972BFF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StringManipulationTest", "tests\unit\StringManipulationTest\StringManipulationTest.csproj", "{17B82D81-DC64-47F1-8BAE-680D1521A259}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StringManipulationTest", "tests\unit\StringManipulationTest\StringManipulationTest.csproj", "{17B82D81-DC64-47F1-8BAE-680D1521A259}"
+EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "CsharpSolution", "CsharpSolution", "{67AB8D35-0211-4BDE-9AEB-85C2F9143AC7}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProjectAGreetingsApp", "src\Assignment4\ProjectAGreetingsApp\ProjectAGreetingsApp.csproj", "{E92D1C37-F76F-4874-9B4F-624AD7CD79B6}"
@@ -69,11 +70,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "XunitTestingAssignment", "t
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CalculationService", "src\Assignment15\CalculationService\CalculationService.csproj", "{85D9CCED-580A-4B34-80C8-2F57321C7142}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Assignment15", "Assignment15", "{72F892C5-888E-4BE2-A51D-63EA94EFA81F}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "XunitTestingAssignment", "tests\unit\Assignment15\XunitTestingAssignment\XunitTestingAssignment.csproj", "{D0D343C6-953F-4261-BBA5-381665826408}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CalculationService", "src\Assignment15\CalculationService\CalculationService.csproj", "{85D9CCED-580A-4B34-80C8-2F57321C7142}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Assignment15", "Assignment15", "{A7A10F56-C469-4653-92F4-1B2D85B22D62}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -106,7 +103,6 @@ Global
 		{D718BC05-D53A-42E4-A378-B9372CD1CA89}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D718BC05-D53A-42E4-A378-B9372CD1CA89}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D718BC05-D53A-42E4-A378-B9372CD1CA89}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D718BC05-D53A-42E4-A378-B9372CD1CA89}.Release|Any CPU.Build.0 = Release|Any CPU
 		{44AD417F-51C5-42AC-A3DC-E5F3F2B80258}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{44AD417F-51C5-42AC-A3DC-E5F3F2B80258}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{44AD417F-51C5-42AC-A3DC-E5F3F2B80258}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -133,6 +129,26 @@ Global
 		{17B82D81-DC64-47F1-8BAE-680D1521A259}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{17B82D81-DC64-47F1-8BAE-680D1521A259}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{17B82D81-DC64-47F1-8BAE-680D1521A259}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E92D1C37-F76F-4874-9B4F-624AD7CD79B6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E92D1C37-F76F-4874-9B4F-624AD7CD79B6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E92D1C37-F76F-4874-9B4F-624AD7CD79B6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E92D1C37-F76F-4874-9B4F-624AD7CD79B6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{638C73DA-FE73-4B4C-8BDC-7F8BDD4DC794}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{638C73DA-FE73-4B4C-8BDC-7F8BDD4DC794}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{638C73DA-FE73-4B4C-8BDC-7F8BDD4DC794}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{638C73DA-FE73-4B4C-8BDC-7F8BDD4DC794}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5FCC09A3-2217-4CA2-B798-54573D8F1C61}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5FCC09A3-2217-4CA2-B798-54573D8F1C61}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5FCC09A3-2217-4CA2-B798-54573D8F1C61}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5FCC09A3-2217-4CA2-B798-54573D8F1C61}.Release|Any CPU.Build.0 = Release|Any CPU
+		{411AC890-50A7-46CA-BF77-BC51BCB7BA25}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{411AC890-50A7-46CA-BF77-BC51BCB7BA25}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{411AC890-50A7-46CA-BF77-BC51BCB7BA25}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{411AC890-50A7-46CA-BF77-BC51BCB7BA25}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9B40DD45-2E17-428D-ACC3-CDEC507A92B6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9B40DD45-2E17-428D-ACC3-CDEC507A92B6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9B40DD45-2E17-428D-ACC3-CDEC507A92B6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9B40DD45-2E17-428D-ACC3-CDEC507A92B6}.Release|Any CPU.Build.0 = Release|Any CPU
 		{D0D343C6-953F-4261-BBA5-381665826408}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D0D343C6-953F-4261-BBA5-381665826408}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D0D343C6-953F-4261-BBA5-381665826408}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -172,9 +188,6 @@ Global
 		{5FCC09A3-2217-4CA2-B798-54573D8F1C61} = {67AB8D35-0211-4BDE-9AEB-85C2F9143AC7}
 		{411AC890-50A7-46CA-BF77-BC51BCB7BA25} = {67AB8D35-0211-4BDE-9AEB-85C2F9143AC7}
 		{9B40DD45-2E17-428D-ACC3-CDEC507A92B6} = {67AB8D35-0211-4BDE-9AEB-85C2F9143AC7}
-		{72F892C5-888E-4BE2-A51D-63EA94EFA81F} = {DCAA095E-2796-41FA-A1D0-1134E29309E7}
-		{D0D343C6-953F-4261-BBA5-381665826408} = {C778E127-C02C-4A37-8E85-3D86DCC063DC}
-		{85D9CCED-580A-4B34-80C8-2F57321C7142} = {72F892C5-888E-4BE2-A51D-63EA94EFA81F}
 		{72F892C5-888E-4BE2-A51D-63EA94EFA81F} = {DCAA095E-2796-41FA-A1D0-1134E29309E7}
 		{D0D343C6-953F-4261-BBA5-381665826408} = {C778E127-C02C-4A37-8E85-3D86DCC063DC}
 		{85D9CCED-580A-4B34-80C8-2F57321C7142} = {72F892C5-888E-4BE2-A51D-63EA94EFA81F}
